2015_12_18_09550 오주석:즉시 구매, 최근본 도서

+++ src/main/java/com/gobook/member/dto/UserGP.java
@@ -3,8 +3,9 @@ package com.gobook.member.dto;
 public class UserGP {
 	private int user_gp_num;
 	private int group_purchase_num;
-	private int book_num;
+	private long book_num;
 	private String member_id;
+	
 	public int getUser_gp_num() {
 		return user_gp_num;
 	}
@@ -17,10 +18,10 @@ public class UserGP {
 	public void setGroup_purchase_num(int group_purchase_num) {
 		this.group_purchase_num = group_purchase_num;
 	}
-	public int getBook_num() {
+	public long getBook_num() {
 		return book_num;
 	}
-	public void setBook_num(int book_num) {
+	public void setBook_num(long book_num) {
 		this.book_num = book_num;
 	}
 	public String getMember_id() {
