+++ src/main/java/com/gobook/userbook/dto/UserBookStarDto.java
@@ -3,7 +3,7 @@ package com.gobook.userbook.dto;
 public class UserBookStarDto {
 	private int userbookstar_num;		//사용자 별점 고유번호
 	private String member_id;			//사용자 아이디
-	private int book_num;				//책 고유번호
+	private long book_num;				//책 고유번호
 	private int userbookstar_star;		//사용자 별점
 	public int getUserbookstar_num() {
 		return userbookstar_num;
@@ -17,10 +17,10 @@ public class UserBookStarDto {
 	public void setMember_id(String member_id) {
 		this.member_id = member_id;
 	}
-	public int getBook_num() {
+	public long getBook_num() {
 		return book_num;
 	}
-	public void setBook_num(int book_num) {
+	public void setBook_num(long book_num) {
 		this.book_num = book_num;
 	}
 	public int getUserbookstar_star() {
